+++ examples/volumetric/densifycloud.py
@@ -1,20 +1,16 @@
-"""Adds new points to an input point cloud.
+"""Generate a denser point cloud.
 The new points are created in such a way that
 all points in any local neighborhood are
-within a target distance of one another.
-"""
+within a target distance of one another"""
 from vedo import *
 import numpy as np
-np.random.seed(3)
 
-npts = 200                       # nr. of points
+npts = 50                        # nr. of points
 coords = np.random.rand(npts, 3) # range is [0, 1]
-scals = np.abs(coords[:, 2])     # let the scalar be the z of point itself
-
+scals = np.abs(coords[:, 1])     # let the scalar be the y of the point itself
 apts = Points(coords, r=9).addPointArray(scals, name='scals')
 
-densecloud = densifyCloud(apts, .05, closestN=10, maxIter=1)
-print('npt increased', apts.N(), '->', densecloud.N())
+densecloud = apts.densify(0.1, closest=10, niter=1) # return a new pointcloud.Points
+printc('nr. points increased', apts.N(), '\rightarrow ', densecloud.N(), c='lg')
 
-ppp = Points(densecloud.points())
-show(apts, densecloud, __doc__, axes=9)
+show([(apts, __doc__), densecloud], N=2, axes=1)